pyproject.toml
[project]
name = "rl-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "gym[classic_control,toy-text]>=0.26.2",
    "numpy>=1.24.4",
    "tqdm>=4.67.0",
    "torch>=2.3.0",
    "fire>=0.6.0",
    "PyYAML>=6.0.1",
    "wandb>=0.17.0",
    "pygame>=2.5.2",
    "scipy>=1.10.0",
    "pettingzoo[mpe]>=1.24.0",
]
